lambda/requirements.txt
snowflake-connector-python==3.11.0